+++ integrations/datadog/tests/conftest.py
@@ -2,6 +2,7 @@ from typing import Generator
 
 import pytest
 from unittest.mock import AsyncMock, MagicMock, patch, PropertyMock
+from datadog.client import DatadogClient
 from port_ocean.context.ocean import initialize_port_ocean_context
 from port_ocean.exceptions.context import PortOceanContextAlreadyInitializedError
 
@@ -45,3 +46,12 @@ def mock_ocean_context(mock_integration_config: dict[str, str]) -> None | MagicM
         # Context already initialized, ignore
         pass
     return None
+
+
+@pytest.fixture
+def mock_datadog_client() -> DatadogClient:
+    return DatadogClient(
+        api_url="https://api.datadoghq.com",
+        api_key="test_api_key",
+        app_key="test_app_key",
+    )
